Move CA1813 AvoidUnsealedAttributes to the new location.
