+++ PortFxCop/src/System.Runtime.Analyzers/UnitTests/AvoidUnsealedAttributesTests.Fixer.cs
@@ -1,33 +1,70 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Test.Utilities;
 using Microsoft.CodeAnalysis.UnitTests;
+using Roslyn.Test.Utilities;
 using Xunit;
 
 namespace System.Runtime.Analyzers.UnitTests
 {
-    public class AvoidUnsealedAttributesFixerTests : CodeFixTestBase
+    public partial class AvoidUnsealedAttributeFixerTests : CodeFixTestBase
     {
         protected override DiagnosticAnalyzer GetBasicDiagnosticAnalyzer()
         {
-            return new BasicAvoidUnsealedAttributesAnalyzer();
+            return new AvoidUnsealedAttributesAnalyzer();
+        }
+                          
+        protected override CodeFixProvider GetBasicCodeFixProvider()
+        {
+            return new AvoidUnsealedAttributesFixer();
         }
 
         protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
         {
-            return new CSharpAvoidUnsealedAttributesAnalyzer();
+            return new AvoidUnsealedAttributesAnalyzer();
         }
-
-        protected override CodeFixProvider GetBasicCodeFixProvider()
-        {
-            return new BasicAvoidUnsealedAttributesFixer();
-        }
-
+                          
         protected override CodeFixProvider GetCSharpCodeFixProvider()
         {
-            return new CSharpAvoidUnsealedAttributesFixer();
+            return new AvoidUnsealedAttributesFixer();
         }
+
+        #region CodeFix Tests
+
+        [Fact]
+        public void CA1813CSharpCodeFixProviderTestFired()
+        {
+            VerifyCSharpFix(@"
+using System;
+
+public class AttributeClass : Attribute
+{
+}", @"
+using System;
+
+public sealed class AttributeClass : Attribute
+{
+}");
+        }
+
+        [Fact]
+        public void CA1813VisualBasicCodeFixProviderTestFired()
+        {
+            VerifyBasicFix(@"
+Imports System
+
+Public Class AttributeClass
+    Inherits Attribute
+End Class", @"
+Imports System
+
+Public NotInheritable Class AttributeClass
+    Inherits Attribute
+End Class");
+        }
+
+        #endregion
     }
 }
