--- PortFxCop/src/System.Runtime.Analyzers/VisualBasic/BasicAvoidUnsealedAttributes.vb
@@ -1,17 +1,0 @@
-' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-Imports Microsoft.CodeAnalysis
-Imports Microsoft.CodeAnalysis.Diagnostics
-Imports Microsoft.CodeAnalysis.VisualBasic
-Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
-
-Namespace System.Runtime.Analyzers   
-    ''' <summary>
-    ''' CA1813: Avoid unsealed attributes
-    ''' </summary>
-    <DiagnosticAnalyzer(LanguageNames.VisualBasic)>
-    Public NotInheritable Class BasicAvoidUnsealedAttributesAnalyzer
-        Inherits AvoidUnsealedAttributesAnalyzer
-
-    End Class
-End Namespace
