--- src/FxCop/System.Runtime.Analyzers/Core/Performance/AvoidUnsealedAttributes.cs
@@ -1,62 +1,0 @@
-﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using System;
-using System.Collections.Immutable;
-using System.Threading;
-using System.Linq;
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis;
-
-namespace System.Runtime.Analyzers
-{
-    /// <summary>
-    /// CA1813: Seal attribute types for improved performance. Sealing attribute types speeds up performance during reflection on custom attributes.
-    /// </summary>
-    [DiagnosticAnalyzer(LanguageNames.CSharp, LanguageNames.VisualBasic)]
-    public sealed class AvoidUnsealedAttributesAnalyzer : DiagnosticAnalyzer
-    {
-        internal const string RuleId = "CA1813";
-        private static readonly LocalizableString s_localizableTitle = new LocalizableResourceString(nameof(SystemRuntimeAnalyzersResources.AvoidUnsealedAttributes), SystemRuntimeAnalyzersResources.ResourceManager, typeof(SystemRuntimeAnalyzersResources));
-        private static readonly LocalizableString s_localizableMessage = new LocalizableResourceString(nameof(SystemRuntimeAnalyzersResources.SealAttributeTypesForImprovedPerf), SystemRuntimeAnalyzersResources.ResourceManager, typeof(SystemRuntimeAnalyzersResources));
-
-        internal static DiagnosticDescriptor Rule = new DiagnosticDescriptor(RuleId,
-                                                                         s_localizableTitle,
-                                                                         s_localizableMessage,
-                                                                         DiagnosticCategory.Performance,
-                                                                         DiagnosticSeverity.Warning,
-                                                                         isEnabledByDefault: false,
-                                                                         helpLinkUri: "http://msdn.microsoft.com/library/ms182267.aspx",
-                                                                         customTags: WellKnownDiagnosticTags.Telemetry);
-
-        public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => ImmutableArray.Create(Rule);
-
-        public override void Initialize(AnalysisContext analysisContext)
-        {
-            analysisContext.RegisterCompilationStartAction(compilationContext =>
-            {
-                var attributeType = WellKnownTypes.Attribute(compilationContext.Compilation);
-                if (attributeType == null)
-                {
-                    return;
-                }
-
-                compilationContext.RegisterSymbolAction(context =>
-                {
-                    AnalyzeSymbol((INamedTypeSymbol)context.Symbol, attributeType, context.ReportDiagnostic);
-                },
-                SymbolKind.NamedType);
-            });
-        }
-
-        private static void AnalyzeSymbol(INamedTypeSymbol namedType, INamedTypeSymbol attributeType, Action<Diagnostic> addDiagnostic)
-        {
-            if (namedType.IsAbstract || namedType.IsSealed || !namedType.GetBaseTypesAndThis().Contains(attributeType))
-            {
-                return;
-            }
-
-            // Non-sealed non-abstract attribute type.
-            addDiagnostic(namedType.CreateDiagnostic(Rule));
-        }
-    }
-}
